# Inductor cpp wrapper: add -ffast-math in linking flag